organization - move some hosted stuff to rust lib, otherwise put the orchestrator code into the appropriate hosted test files
