+++ libraries/rust/src/lib.rs
@@ -53,12 +53,16 @@
 
 /// Instruction builders for programs and adapters supported by the SDK
 pub mod ix_builder;
+/// things that should be provided by the solana sdk, but are not
+pub mod solana;
 /// Utilities for swap adapters
 pub mod swap;
 /// Utilities for tokens and token prices
 pub mod tokens;
 /// Transaction builder
 pub mod tx_builder;
+/// General purpose logic used by this lib and clients, unrelated to jet or solana
+pub mod util;
 
 pub use jet_control;
 pub use jet_margin;
