+++ libraries/rust/src/util/mod.rs
@@ -1,0 +1,4 @@
+/// simplify parallel execution of generic tasks
+pub mod asynchronous;
+/// generic processing of arbitrary data
+pub mod data;
