+++ tests/hosted/src/pricing.rs
@@ -1,25 +1,23 @@
-use std::collections::hash_map::Entry;
 use std::collections::{HashMap, HashSet};
 use std::sync::Arc;
 
 use anyhow::Result;
 
 use itertools::Itertools;
+use jet_margin_sdk::cat;
+use jet_margin_sdk::solana::keypair::clone;
+use jet_margin_sdk::solana::transaction::{SendTransactionBuilder, TransactionBuilder};
 use jet_margin_sdk::swap::SwapPool;
 use jet_margin_sdk::tokens::TokenPrice;
+use jet_margin_sdk::util::asynchronous::{AndAsync, MapAsync};
 use jet_simulation::solana_rpc_api::SolanaRpcClient;
-use jet_static_program_registry::orca_swap_v2;
 use solana_sdk::pubkey::Pubkey;
-use solana_sdk::signature::{Keypair, Signer};
+use solana_sdk::signature::Keypair;
 
-use jet_margin_pool::{Amount, TokenChange};
 use tokio::try_join;
 
-use crate::context::MarginTestContext;
-use crate::margin::MarginUser;
-use crate::swap::SwapPoolConfig;
+use crate::swap::{SwapPoolConfig, SwapRegistry};
 use crate::tokens::TokenManager;
-use crate::{cat, clone, AndAsync, Concat, MapAsync, SendTransactionBuilder, TransactionBuilder};
 
 pub const ONE: u64 = 1_000_000_000;
 
@@ -41,45 +39,6 @@ impl Clone for TokenPricer {
             swap_registry: self.swap_registry.clone(),
         }
     }
-}
-
-pub type SwapRegistry = HashMap<Pubkey, HashMap<Pubkey, SwapPool>>;
-
-pub async fn create_swap_pools(
-    rpc: &Arc<dyn SolanaRpcClient>,
-    mints: &[Pubkey],
-) -> Result<SwapRegistry> {
-    let mut registry = SwapRegistry::new();
-    for (one, two, pool) in mints
-        .iter()
-        .combinations(2)
-        .map(|c| (*c[0], *c[1]))
-        .map_async(|(one, two)| create_and_insert(rpc, one, two))
-        .await?
-    {
-        registry.entry(one).or_default().insert(two, pool);
-        registry.entry(two).or_default().insert(one, pool);
-    }
-
-    Ok(registry)
-}
-
-async fn create_and_insert(
-    rpc: &Arc<dyn SolanaRpcClient>,
-    one: Pubkey,
-    two: Pubkey,
-) -> Result<(Pubkey, Pubkey, SwapPool)> {
-    let pool = SwapPool::configure(
-        rpc,
-        &orca_swap_v2::id(),
-        &one,
-        &two,
-        100_000_000 * ONE,
-        100_000_000 * ONE,
-    )
-    .await?;
-
-    Ok((one, two, pool))
 }
 
 impl TokenPricer {
@@ -349,179 +308,4 @@ fn set_constant_product_price(
 
         move_a = sqrt(product/desired_price) - balance_a
     */
-}
-
-/// A MarginUser that takes some extra liberties
-#[derive(Clone)]
-pub struct TestUser<'a> {
-    pub ctx: &'a MarginTestContext,
-    pub user: MarginUser,
-    pub mint_to_token_account: HashMap<Pubkey, Pubkey>,
-}
-
-impl<'a> std::fmt::Debug for TestUser<'a> {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("TestUser")
-            .field("user", &self.user.address())
-            // .field("liquidator", &self.liquidator.address())
-            .field("mint_to_token_account", &self.mint_to_token_account)
-            .finish()
-    }
-}
-
-impl<'a> TestUser<'a> {
-    pub async fn token_account(&mut self, mint: &Pubkey) -> Result<Pubkey> {
-        let token_account = match self.mint_to_token_account.entry(*mint) {
-            Entry::Occupied(entry) => *entry.get(),
-            Entry::Vacant(entry) => *entry.insert(
-                self.ctx
-                    .tokens
-                    .create_account(mint, self.user.owner())
-                    .await?,
-            ),
-        };
-
-        Ok(token_account)
-    }
-
-    pub async fn ephemeral_token_account(&self, mint: &Pubkey, amount: u64) -> Result<Pubkey> {
-        self.ctx
-            .tokens
-            .create_account_funded(mint, self.user.owner(), amount)
-            .await
-    }
-
-    pub async fn mint(&mut self, mint: &Pubkey, amount: u64) -> Result<()> {
-        let token_account = self.token_account(mint).await?;
-        self.ctx.tokens.mint(mint, &token_account, amount).await
-    }
-
-    pub async fn deposit(&self, mint: &Pubkey, amount: u64) -> Result<()> {
-        let token_account = self.ephemeral_token_account(mint, amount).await?;
-        self.user
-            .deposit(mint, &token_account, TokenChange::shift(amount))
-            .await?;
-        self.ctx.tokens.refresh_to_same_price(mint).await
-    }
-
-    pub async fn deposit_from_wallet(&mut self, mint: &Pubkey, amount: u64) -> Result<()> {
-        let token_account = self.token_account(mint).await?;
-        self.user
-            .deposit(mint, &token_account, TokenChange::shift(amount))
-            .await
-    }
-
-    pub async fn borrow(&self, mint: &Pubkey, amount: u64) -> Result<()> {
-        self.ctx.tokens.refresh_to_same_price(mint).await?;
-        self.user.refresh_all_pool_positions().await?;
-        self.user.borrow(mint, TokenChange::shift(amount)).await
-    }
-
-    pub async fn borrow_to_wallet(&self, mint: &Pubkey, amount: u64) -> Result<()> {
-        self.borrow(mint, amount).await?;
-        self.withdraw(mint, amount).await
-    }
-
-    pub async fn margin_repay(&self, mint: &Pubkey, amount: u64) -> Result<()> {
-        self.user
-            .margin_repay(mint, TokenChange::shift(amount))
-            .await
-    }
-
-    pub async fn withdraw(&self, mint: &Pubkey, amount: u64) -> Result<()> {
-        let token_account = self.ephemeral_token_account(mint, 0).await?;
-        self.user.refresh_all_pool_positions().await?;
-        self.user
-            .withdraw(mint, &token_account, TokenChange::shift(amount))
-            .await
-    }
-
-    pub async fn withdraw_to_wallet(&mut self, mint: &Pubkey, amount: u64) -> Result<()> {
-        let token_account = self.token_account(mint).await?;
-        self.user.refresh_all_pool_positions().await?;
-        self.user
-            .withdraw(mint, &token_account, TokenChange::shift(amount))
-            .await
-    }
-
-    pub async fn swap(
-        &self,
-        swaps: &SwapRegistry,
-        src: &Pubkey,
-        dst: &Pubkey,
-        amount: u64,
-    ) -> Result<()> {
-        let pool = swaps.get(src).unwrap().get(dst).unwrap();
-        let transit_src = self
-            .ctx
-            .tokens
-            .create_account(src, self.user.address())
-            .await?;
-        let transit_dst = self
-            .ctx
-            .tokens
-            .create_account(dst, self.user.address())
-            .await?;
-        self.user
-            .swap(
-                &orca_swap_v2::id(),
-                src,
-                dst,
-                &transit_src,
-                &transit_dst,
-                pool,
-                Amount::tokens(amount),
-                Amount::tokens(0),
-            )
-            .await
-    }
-
-    pub async fn liquidate_end(&self, liquidator: Pubkey) -> Result<()> {
-        self.user.liquidate_end(Some(liquidator)).await
-    }
-}
-
-#[derive(Debug)]
-pub struct TestLiquidator<'a> {
-    pub ctx: &'a MarginTestContext,
-    pub wallet: Keypair,
-}
-
-impl<'a> TestLiquidator<'a> {
-    pub async fn new(ctx: &'a MarginTestContext) -> Result<TestLiquidator> {
-        Ok(TestLiquidator {
-            ctx,
-            wallet: ctx.create_liquidator(100).await?,
-        })
-    }
-
-    pub async fn begin(&self, user: &MarginUser) -> Result<TestUser<'a>> {
-        let liquidation = self
-            .ctx
-            .margin
-            .liquidator(&self.wallet, user.owner(), user.seed())
-            .await?;
-        liquidation.liquidate_begin(true).await?;
-
-        Ok(TestUser {
-            ctx: self.ctx,
-            user: liquidation,
-            mint_to_token_account: HashMap::new(),
-        })
-    }
-
-    pub async fn liquidate(
-        &self,
-        user: &MarginUser,
-        swaps: &SwapRegistry,
-        collateral: &Pubkey,
-        sell: u64,
-        loan: &Pubkey,
-        repay: u64,
-    ) -> Result<()> {
-        let liq = self.begin(user).await?;
-        liq.swap(swaps, collateral, loan, sell).await?;
-        liq.margin_repay(loan, repay).await?;
-        liq.liquidate_end(self.wallet.pubkey()).await
-    }
 }
